Renamed some files for more clarity
